feat: mocked sqlite db

+++ src/models/user.rs
@@ -21,13 +21,13 @@ pub struct User {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct UserBD {
-    pub id: i32,
+    pub id: i64,
     pub full_name: String,
     pub profile_pic: String,
     pub cpf: String,
     pub birthdate: String, // Use a string in ISO 8601 format
     pub marital_status: String,
-    pub gross_mensal_income: i32,
+    pub gross_mensal_income: i64,
     pub email: String,
     pub phone_number: String,
     pub is_admin: bool,
@@ -35,12 +35,12 @@ pub struct UserBD {
     pub user_password: String,
     pub created_at: String,
     pub updated_at: String,
-    pub address_id: i32, // Assuming you have the address ID available
+    pub address_id: i64, // Assuming you have the address ID available
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Address {
-    pub id: i32,
+    pub id: i64,
     pub zip_code: String,
     pub city: String,
     pub state: String,
